src/requirements.txt
django
djangorestframework
django-mptt
djoser
python-dotenv
djangorestframework-recursive
psycopg2-binary
drf-yasg